# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: LedgerLite API - Financial management for SMEs
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List all transactions
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [income, expense]
        - name: category
          in: query
          schema:
            type: string
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a new transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionBody"
      responses:
        "201":
          description: Created transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a transaction by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "404":
          description: Not found
    put:
      operationId: updateTransaction
      tags: [transactions]
      summary: Update a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTransactionBody"
      responses:
        "200":
          description: Updated transaction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
    delete:
      operationId: deleteTransaction
      tags: [transactions]
      summary: Delete a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Customers
...
  /reports/profit-loss:
    get:
      operationId: getProfitLossReport
      tags: [reports]
      summary: Get profit and loss report
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: P&L report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfitLossReport"
  /reports/expense-breakdown:
    get:
      operationId: getExpenseBreakdown
      tags: [reports]
      summary: Get expense breakdown by category
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Expense breakdown by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryBreakdown"
  /reports/revenue-breakdown:
    get:
      operationId: getRevenueBreakdown
      tags: [reports]
      summary: Get revenue breakdown by category
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Revenue breakdown by category
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CategoryBreakdown"
components:
  schemas:
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
        category:
          type: string
        description:
          type: string
        paymentMethod:
          type: string
          enum: [cash, bank, pos, transfer]
        date:
          type: string
          format: date-time
        customerId:
          type: integer
          nullable: true
        customerName:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, type, amount, category, paymentMethod, date, createdAt, updatedAt]
    TransactionList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [data, total, page, limit]
    CreateTransactionBody:
      type: object
      properties:
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
        category:
          type: string
        description:
          type: string
        paymentMethod:
          type: string
          enum: [cash, bank, pos, transfer]
        date:
          type: string
          format: date-time
        customerId:
          type: integer
          nullable: true
      required: [type, amount, category, paymentMethod, date]
    UpdateTransactionBody:
      type: object
      properties:
        type:
          type: string
          enum: [income, expense]
        amount:
          type: number
        category:
          type: string
        description:
          type: string
        paymentMethod:
          type: string
          enum: [cash, bank, pos, transfer]
        date:
          type: string
          format: date-time
        customerId:
          type: integer
          nullable: true

    # Customer schemas
...
    ProfitLossReport:
      type: object
      properties:
        totalRevenue:
          type: number
        totalExpenses:
          type: number
        grossProfit:
          type: number
        profitMargin:
          type: number
        revenueByCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryBreakdown"
        expenseByCategory:
          type: array
          items:
            $ref: "#/components/schemas/CategoryBreakdown"
        startDate:
          type: string
        endDate:
          type: string
      required: [totalRevenue, totalExpenses, grossProfit, profitMargin, revenueByCategory, expenseByCategory, startDate, endDate]
    CategoryBreakdown:
      type: object
      properties:
        category:
          type: string
        amount:
          type: number
        percentage:
          type: number
        count:
          type: integer
      required: [category, amount, percentage, count]